backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
djangorestframework==3.16.1
sqlparse==0.5.3
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.3.1
